feat #110 - Gestion de la déconnexion

+++ client/src/components/NavBar/NavBar.js
@@ -218,9 +218,45 @@ export default function NavBar(props) {
   }
 
   function handleLogout() {
-    localStorage.clear();
-    setIsConnected(false);
-  }
+    const url = api + 'users/logout';
+    const userJson = {
+      username: user,
+      password: password,
+    };
+    console.log(JSON.stringify(userJson));
+    fetch(url, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify(userJson)
+    })
+      .then(response => response.json())
+      .then(responseFromServer => {
+        console.log(responseFromServer);
+        if (responseFromServer.status === 401) {
+          Swal.fire({
+            icon: 'error',
+            title: 'Oops...',
+            text: responseFromServer.detail,
+          });
+        }
+        else {
+          Swal.fire({
+            position: 'top-end',
+            icon: 'success',
+            title: userJson.username + ' Dscconnected !',
+            showConfirmButton: false,
+            timer: 1500
+          })
+          localStorage.clear();
+          setIsConnected(false);
+        }
+      })
+      .catch((error) => {
+        console.log(error);
+      })
+    }
 
   function handleRegister(e) {
     const url = api + 'users';
